requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.9
django-allauth==0.46.0
djangorestframework==3.12.4
drf-yasg==1.20.0
fasttext==0.9.2
graphviz==0.8.4
idna==2.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mxnet-cu92==1.5.0
numpy==1.16.6
oauthlib==3.1.1
packaging==21.3
pybind11==2.8.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.18.4
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.22
